Added other pages and finished rebuild of site.
Added the Services, About Us, and Contact Us pages and the information for
them and updated all of the linkes for them and the broken squirrel link
on the sidebar so the image actually links(these still do not actually
link to anything). This is the finished version of the responsive rebuild
of the site as it is currently.